+++ src/components/layouts/Footer.tsx
@@ -1,0 +1,3 @@
+export const Footer = () => {
+  return <div>フッター</div>;
+};
