+++ src/components/layouts/Header.tsx
@@ -1,0 +1,3 @@
+export const Header = () => {
+  return <div>ヘッダー</div>;
+};
